Move TrainValidation to file and update controlflow
